Moves promotedOffers module to theme. Merges promotedOffers store into one file.
