--- src/modules/promoted-offers/store/index.ts
@@ -1,21 +1,0 @@
-import { Module } from 'vuex'
-import getters from './getters'
-import actions from './actions'
-import mutations from './mutations'
-import RootState from '@vue-storefront/core/types/RootState'
-import PromotedOffersState from '../types/PromotedOffersState'
-
-export const module: Module<PromotedOffersState, RootState> = {
-  namespaced: true,
-  state: {
-    banners: {
-      mainBanners: [],
-      smallBanners: [],
-      productBanners: []
-    },
-    headImage: null
-  },
-  getters,
-  actions,
-  mutations
-}
