+++ src/themes/default/store/index.ts
@@ -1,6 +1,8 @@
 import { VueStorefrontModule } from '@vue-storefront/core/lib/module';
 import { Homepage } from './homepage'
+import { PromotedOffers } from './promoted-offers'
 
 export const registerModules: VueStorefrontModule[] = [
   Homepage,
+  PromotedOffers
 ];
