+++ src/themes/default/store/promoted-offers/types/PromotedOffersState.ts
